conf/requirements.txt
python-dotenv
mysql-connector-python
requests
sshtunnel
